Move demographic models to separate folder
