+++ src/main/java/org/sagebionetworks/bridge/hibernate/HibernateDemographicDao.java
@@ -9,8 +9,8 @@ import org.sagebionetworks.bridge.dao.DemographicDao;
 import org.sagebionetworks.bridge.hibernate.QueryBuilder.WhereClauseBuilder;
 import org.sagebionetworks.bridge.models.PagedResourceList;
 import org.sagebionetworks.bridge.models.SearchTermPredicate;
-import org.sagebionetworks.bridge.models.studies.Demographic;
-import org.sagebionetworks.bridge.models.studies.DemographicUser;
+import org.sagebionetworks.bridge.models.demographics.Demographic;
+import org.sagebionetworks.bridge.models.demographics.DemographicUser;
 import org.springframework.stereotype.Component;
 
 /**
