+++ src/main/java/org/sagebionetworks/bridge/models/accounts/ParticipantVersion.java
@@ -9,7 +9,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import org.sagebionetworks.bridge.dynamodb.DynamoParticipantVersion;
 import org.sagebionetworks.bridge.json.BridgeTypeName;
 import org.sagebionetworks.bridge.models.BridgeEntity;
-import org.sagebionetworks.bridge.models.studies.Demographic;
+import org.sagebionetworks.bridge.models.demographics.Demographic;
 
 /**
  * Represents a de-identified snapshot of a study participant at a moment in time. This is used by Exporter 3.0 to
