+++ src/main/java/org/sagebionetworks/bridge/validators/DemographicValidator.java
@@ -1,8 +1,8 @@
 package org.sagebionetworks.bridge.validators;
 
 import org.apache.commons.lang3.StringUtils;
-import org.sagebionetworks.bridge.models.studies.Demographic;
-import org.sagebionetworks.bridge.models.studies.DemographicValue;
+import org.sagebionetworks.bridge.models.demographics.Demographic;
+import org.sagebionetworks.bridge.models.demographics.DemographicValue;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
