+++ src/main/java/org/sagebionetworks/bridge/validators/DemographicValueValidator.java
@@ -4,7 +4,7 @@ import static org.sagebionetworks.bridge.validators.Validate.CANNOT_BE_NULL_OR_E
 import static org.sagebionetworks.bridge.validators.ValidatorUtils.validateStringLength;
 
 import org.apache.commons.lang3.StringUtils;
-import org.sagebionetworks.bridge.models.studies.DemographicValue;
+import org.sagebionetworks.bridge.models.demographics.DemographicValue;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
