+++ src/main/java/org/sagebionetworks/bridge/validators/DemographicValuesValidationConfigValidator.java
@@ -4,7 +4,7 @@ import static org.sagebionetworks.bridge.validators.Validate.CANNOT_BE_NULL;
 
 import java.io.IOException;
 
-import org.sagebionetworks.bridge.models.studies.DemographicValuesValidationConfig;
+import org.sagebionetworks.bridge.models.demographics.DemographicValuesValidationConfig;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
