+++ src/test/java/org/sagebionetworks/bridge/json/DemographicValueDeserializerTest.java
@@ -3,7 +3,7 @@ package org.sagebionetworks.bridge.json;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNull;
 
-import org.sagebionetworks.bridge.models.studies.DemographicValue;
+import org.sagebionetworks.bridge.models.demographics.DemographicValue;
 import org.testng.annotations.Test;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
