+++ src/test/java/org/sagebionetworks/bridge/validators/DemographicValuesValidatorTest.java
@@ -7,10 +7,10 @@ import java.io.IOException;
 
 import org.sagebionetworks.bridge.exceptions.InvalidEntityException;
 import org.sagebionetworks.bridge.json.BridgeObjectMapper;
-import org.sagebionetworks.bridge.models.studies.Demographic;
-import org.sagebionetworks.bridge.models.studies.DemographicUser;
-import org.sagebionetworks.bridge.models.studies.DemographicValue;
-import org.sagebionetworks.bridge.models.studies.DemographicValuesValidationConfig;
+import org.sagebionetworks.bridge.models.demographics.Demographic;
+import org.sagebionetworks.bridge.models.demographics.DemographicUser;
+import org.sagebionetworks.bridge.models.demographics.DemographicValue;
+import org.sagebionetworks.bridge.models.demographics.DemographicValuesValidationConfig;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
